// src/Organizer.SPA/src/app/api/OrganizerApiClient/model/utilizationRow.ts
/**
 *  Organizer
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UtilizationRow { 
    monthNumber?: number;
    description?: string;
    billableHours?: number;
    toBeBilledHours?: number;
    billedHours?: number;
    readonly percentageBilledHours?: number;
    target?: number;
    daysToTarget?: number;
}